LPS-56522 Move AccessControlUtil into portal-service
