+++ portal-impl/src/META-INF/util-spring.xml
@@ -529,7 +529,7 @@
 	</bean>
 	<bean id="com.liferay.portal.sanitizer.DummySanitizerImpl" class="com.liferay.portal.sanitizer.DummySanitizerImpl" />
 	<bean id="com.liferay.portal.scripting.ruby.RubyExecutor" class="com.liferay.portal.scripting.ruby.RubyExecutor" depends-on="com.liferay.portal.kernel.util.FileUtil" />
-	<bean id="com.liferay.portal.security.access.control.AccessControlUtil" class="com.liferay.portal.security.access.control.AccessControlUtil">
+	<bean id="com.liferay.portal.kernel.security.access.control.AccessControlUtil" class="com.liferay.portal.kernel.security.access.control.AccessControlUtil">
 		<property name="accessControl">
 			<bean class="com.liferay.portal.security.access.control.AccessControlImpl" />
 		</property>
