+++ portal-impl/src/com/liferay/portal/security/access/control/AllowedHostsAccessControlPolicy.java
@@ -14,6 +14,7 @@
 
 package com.liferay.portal.security.access.control;
 
+import com.liferay.portal.kernel.security.access.control.AccessControlUtil;
 import com.liferay.portal.kernel.security.access.control.AccessControlled;
 import com.liferay.portal.kernel.security.access.control.BaseAccessControlPolicy;
 import com.liferay.portal.kernel.util.MapUtil;
