+++ portal-impl/src/com/liferay/portal/servlet/AtomServlet.java
@@ -19,9 +19,9 @@ import com.liferay.portal.atom.AtomUtil;
 import com.liferay.portal.kernel.atom.AtomCollectionAdapter;
 import com.liferay.portal.kernel.atom.AtomCollectionAdapterRegistryUtil;
 import com.liferay.portal.kernel.security.access.control.AccessControlThreadLocal;
+import com.liferay.portal.kernel.security.access.control.AccessControlUtil;
 import com.liferay.portal.kernel.security.auth.verifier.AuthVerifierResult;
 import com.liferay.portal.model.User;
-import com.liferay.portal.security.access.control.AccessControlUtil;
 import com.liferay.portal.security.auth.AccessControlContext;
 import com.liferay.portal.service.UserLocalServiceUtil;
