+++ portal-service/src/com/liferay/portal/kernel/security/access/control/AccessControlUtil.java
@@ -12,11 +12,11 @@
  * details.
  */
 
-package com.liferay.portal.security.access.control;
+package com.liferay.portal.kernel.security.access.control;
 
 import com.liferay.portal.kernel.exception.PortalException;
-import com.liferay.portal.kernel.security.access.control.AccessControl;
 import com.liferay.portal.kernel.security.auth.verifier.AuthVerifierResult;
+import com.liferay.portal.kernel.security.pacl.permission.PortalRuntimePermission;
 import com.liferay.portal.kernel.util.AutoResetThreadLocal;
 import com.liferay.portal.security.auth.AccessControlContext;
 import com.liferay.portal.security.auth.AuthException;
@@ -34,14 +34,15 @@ import javax.servlet.http.HttpServletResponse;
 public class AccessControlUtil {
 
 	public static AccessControl getAccessControl() {
-		if (_accessControl == null) {
-			_accessControl = new AccessControlImpl();
-		}
+		PortalRuntimePermission.checkGetBeanProperty(AccessControlUtil.class);
 
 		return _accessControl;
 	}
 
 	public static AccessControlContext getAccessControlContext() {
+		PortalRuntimePermission.checkGetBeanProperty(
+			AccessControlUtil.class, "accessControlContext");
+
 		return _accessControlContext.get();
 	}
 
@@ -60,6 +61,9 @@ public class AccessControlUtil {
 	public static void setAccessControlContext(
 		AccessControlContext accessControlContext) {
 
+		PortalRuntimePermission.checkSetBeanProperty(
+			AccessControlUtil.class, "accessControlContext");
+
 		_accessControlContext.set(accessControlContext);
 	}
 
@@ -70,6 +74,8 @@ public class AccessControlUtil {
 	}
 
 	public void setAccessControl(AccessControl accessControl) {
+		PortalRuntimePermission.checkSetBeanProperty(getClass());
+
 		_accessControl = accessControl;
 	}
